Smart Contract and SBT Token/offchain/requirements.txt
fastapi
uvicorn
web3
eth-account
requests
python-dotenv
python-dotenv
pydantic
